% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_and_col.R
\docType{data}
\name{colors_UChicago}
\alias{colors_UChicago}
\title{Set of colors for the University of Chicago}
\format{An object of class \code{character} of length 25.}
\usage{
colors_UChicago[name]
}
\value{
Named characters of the hex code
}
\description{
Simply a set of hex codes and names for colors I like
}
\keyword{datasets}
