% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_and_col.R
\name{get_UChicago_colors}
\alias{get_UChicago_colors}
\title{Retrieve colors from University of Chicago colors}
\usage{
get_UChicago_colors(color1, color2)
}
\value{
a set of named characters
}
\description{
A function to retrieve a list of colors from the list of named colors in colors_UChicago
}
\examples{
get_UChicago_colors('maroon', 'dark gray' , 'light gray' )
}
