% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_and_col.R
\name{get_UChicago_palettes}
\alias{get_UChicago_palettes}
\title{Retrieve Delabj color palettes}
\usage{
get_UChicago_palettes(palette = "core", reverse = FALSE, ...)
}
\description{
Easily retrieve various palettes from the delabj collection.
}
