% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_and_col.R
\docType{data}
\name{palettes_UChicago}
\alias{palettes_UChicago}
\title{University of Chicago Color Pallets}
\format{An object of class \code{list} of length 5.}
\usage{
palettes_UChicago[pallet_name]
}
\description{
A collection of pallets based on the University of Chicago's style guidlines
}
\keyword{datasets}
