% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_UChicago.R
\name{scale_color_UChicago}
\alias{scale_color_UChicago}
\title{Change color to one of based on the University of Chicago's palettes}
\usage{
scale_color_UChicago(palette="main", reverse=FALSE, ...)
}
\arguments{
\item{palette}{Name of the palette ("core", "shade", "tint")}

\item{reverse}{Boolean: should the palette be reversed}
}
\description{
Custom color scale to use the delabj palettes.
}
